requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
google-generativeai==0.5.4
google-cloud-aiplatform==1.56.0
google-cloud-speech==2.26.1
pydub==0.25.1
SQLAlchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
requests==2.32.3
httpx==0.27.0 